+++ dungen/models/__init__.py
@@ -1,0 +1,4 @@
+from .data_model import Player, EncounterEntry
+from .config import Config
+
+__all__ = ['Player', 'EncounterEntry', 'Config']
