+++ dungen/ui/__init__.py
@@ -1,0 +1,3 @@
+from .panels import Panels
+
+__all__ = ['Panels']
